Watchlist hooks, db methods
With the hooks infrastructure in place on the ept server, this module is
the first to take advantage of it and moves all the 'watching' based
methods from boards/threads/posts into this module rather than relying
on the other modules to support this feature.

This includes moving the boards.watching and threads.watching db methods
over to this module, and the updating the posts.byThread and
threads.byBoard method to handle the hooks infra such that the hooks
established in this module are used.

The hooks established in this commit basically check if the board or
thread being referred to in a route is being watched by the user or not.
This data is then placed at the correct position in the route iteself.